Refactor Resque_Worker to use reservers.
Resque_Worker now takes an instance of ReserverInterface as a dependency.

The body of the Resque_Worker::reserve() method has been replaced
with a call to ReserverInterface::reserve().

Blocking and non-blocking reservation behaviour is now implemented
in BlockingListPopReserver and QueueOrderReserver respectively.

Logic that decides whether the worker should sleep
after an attempt to reserve a job has been replaced with a call
to ReserverInterface::waitAfterReservationAttempt().